+++ fabric/src/main/java/com/lx862/mtrscripting/core/util/sound/NonPositionedSoundCall.java
@@ -5,12 +5,12 @@ import org.mtr.mapping.holder.*;
 import org.mtr.mapping.mapper.SoundHelper;
 
 public class NonPositionedSoundCall extends SoundCall {
-    public NonPositionedSoundCall(Identifier id, float volume, float pitch) {
-        super(SoundHelper.createSoundEvent(id), volume, pitch);
+    public NonPositionedSoundCall(Identifier id, SoundCategory soundCategory, float volume, float pitch) {
+        super(SoundHelper.createSoundEvent(id), soundCategory, volume, pitch);
     }
 
     @Override
     public void run(World world, ScriptVector3f basePos) {
-        MinecraftClient.getInstance().getSoundManager().play(new SoundInstance(new NonPositionedSoundInstance(soundEvent, SoundCategory.valueOf(soundCategory), volume, pitch)));
+        MinecraftClient.getInstance().getSoundManager().play(new SoundInstance(new NonPositionedSoundInstance(soundEvent, soundCategory, volume, pitch)));
     }
 }