+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add `MinecraftClientWrapper.getRedstoneLevel(pos: BlockPos)` to obtain the emitted redstone level in the world at a specific BlockPos.
 - Add `Resources.hasSystemFont()` to check if a font is available in the system
 - `EyecandyBlockWrapper.redstoneLevel()` now returns the true redstone level, instead of only 0 and 15.
+- Add an extra `soundCategory: String` argument override for `SoundManager.playSound` and `SoundManager.playLocalSound` to play sound other than in the "Master" category.
 - **PIDS Scripting:**
   - Add `TextWrapper.renderType` to specify a custom render type for text, alongside TextureWrapper.
   - Add `TextWrapper.naturalLight()` and `TextureWrapper.naturalLight()` to allow rendering PIDS elements at world light.